# Base transaction wrapper for updating positions