@TeatroIO 2015-01-04
I've prepared a stage. <a href='http://benmorganio-base-transaction-wrapper-for-option-positions.spree-spree-9ed724b77adec04c49c2.ttrcloud.com' target='_blank'>Click to open.</a>